builtin.lua: Add LspInlayHint
The support for LspInlayHint was added i Neovim v0.10.0 and some
language plugins only support this feature for inlay hints like
rustaceanvim. This results in inlay hints with a black foreground color
on a black background color which is quite hard to interpret ;).

This patch will add support for LspInlayHint and set the color to the
midnight's hint color.